Adding builders to the pod. Creating target for objc code.
Adding the builders to the LayoutKitObjC pod.
This required moving the objc enums to the header files, since objc builders cannot depend on Swift implementation.
Since the cross-language integration depends on header files that match the module name, creating a new LayoutKitObjC target for Xcode build to use to match the cocoa pods build.
Verified that build succeeds for Xcode and `pod lib lint`.