Reset reused views' transform and layer anchor point (#231)
Here's the current order of relevant operations in makeViews:
1) a view gets created or recycled
2) its frame gets set
3) the view gets configured by the layout and config block

Before this change, the recycled view may have had a non-default transform. In addition, the configuration in step 3 may change the transform as well. When setting transform and frame, the order of these operations affects the resulting frame, therefore it's important to enforce a consistent order. With this commit, LayoutKit implementation will be setting the transform to the identity transform when a view is recycled, so that that the view transform and frame are always in the same state at the end of step 2 when control transfers from the core LayoutKit view creation code to any custom view config code.

This way, if a config block sets a transform on the view, it will actually affect the view. Before this change, that setter may or may not change the frame, depending on what transform the view happened to have when it got recycled.